+++ kibot/config_reader.py
@@ -379,6 +379,7 @@ class CfgYamlReader(object):
         GS.global_output = GS.global_from_cli.get('output', None)
         GS.global_dir = GS.global_from_cli.get('dir', None)
         GS.global_variant = GS.global_from_cli.get('variant', None)
+        GS.global_units = GS.global_from_cli.get('units', None)
         GS.global_date_time_format = GS.global_from_cli.get('date_time_format', None)
         GS.global_date_format = GS.global_from_cli.get('date_format', None)
         GS.global_time_format = GS.global_from_cli.get('time_format', None)